Fix: 소수점 주식 매매 기능 수정 및 체결 테스트

+++ trading-service/src/main/java/com/example/trading_service/domain/HoldingCache.java
@@ -26,7 +26,7 @@ public class HoldingCache {
     private Stock stock;
 
     @Column(name = "quantity", nullable = false)
-    private int quantity; // 현재 보유 수량
+    private float quantity; // 현재 보유 수량
 
     @Column(name = "avg_cost", nullable = false)
     private float avgCost; // 평균 매입 단가
