+++ trading-service/src/main/java/com/example/trading_service/service/TradingService.java
@@ -460,7 +460,7 @@ public class TradingService {
     }
 
     // 보유 종목 업데이트
-    private void updateHolding(UUID accountId, UUID stockId, int quantity, float price, boolean isBuy) {
+    private void updateHolding(UUID accountId, UUID stockId, float quantity, float price, boolean isBuy) {
         Optional<HoldingCache> existingHolding = holdingCacheRepository
                 .findByAccountIdAndStockId(accountId, stockId);
         
@@ -591,7 +591,7 @@ public class TradingService {
                 stock.getStockCode() != null ? stock.getStockCode() : "",
                 stock.getStockName() != null ? stock.getStockName() : "",
                 order.getOrderType().toString(),
-                (int) trade.getQuantity(),
+                trade.getQuantity(),
                 trade.getPrice(),
                 trade.getCreatedAt(),
                 "FILLED"
@@ -806,7 +806,7 @@ public class TradingService {
                 stock.getStockCode() != null ? stock.getStockCode() : "",
                 stock.getStockName() != null ? stock.getStockName() : "",
                 order.getOrderType().toString(),
-                (int) order.getQuantity(),
+                order.getQuantity(),
                 order.getPrice(),
                 order.getStatus().toString(),
                 order.getCreatedAt(),
